Rename types/raycast.d.ts to types/iterators.d.ts
